requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.12.14
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime
antlr4-tools==0.2.1
anyio==4.7.0
attrs==24.2.0
bidict==0.23.1
bitarray==3.0.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
dill==0.3.8
docker-pycreds==0.4.0
docutils==0.21.2
effdet==0.3.0
elastic-transport==8.15.1
elasticsearch==8.17.1
execnet==2.1.1
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
get-annotations==0.1.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
grobid-client-python==0.0.5
h11==0.16.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub
idna==3.10
ifaddr==0.2.0
iniconfig==2.0.0
install-jdk==1.1.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
latexcodec==3.0.0
layoutparser==0.3.4
lightning-utilities==0.14.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.15.0
mypy-extensions==1.0.0
ncls==0.0.68
necessary==0.4.3
networkx==3.4.2
nicegui==2.9.1
ninja==1.11.1.2
numpy==2.1.3
omegaconf
opencv-python==4.11.0.86
orjson==3.10.12
packaging==24.2
pandas==2.2.3
papermage==0.20.0
pathvalidate==3.2.3
pdf2image==1.17.0
pdfminer.six==20220524
pdfplumber==0.7.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.5
pscript==0.7.7
psutil==7.0.0
pyarrow==18.1.0
pybtex==0.24.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.1
pysbd==0.3.4
pytest==8.3.5
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.19
python-socketio==5.11.4
pytorch-lightning==2.5.0.post0
pytz==2024.2
pywin32==309; platform_system == 'Windows'
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.4
requirements-parser==0.11.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers
sentry-sdk==2.22.0
seqeval==1.2.2
setproctitle==1.3.5
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.1.0
smashed==0.1.10
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
springs==1.13.0
srsly==2.5.1
starlette==0.47.2
sympy==1.13.1
thefuzz==0.22.1
thinc==8.3.4
threadpoolctl==3.5.0
timm==1.0.15
tokenizers
torch
torchaudio
torchmetrics
torchvision
tqdm==4.67.0
transformers
typer==0.15.2
types-PyYAML==6.0.12.20241230
types-setuptools==76.0.0.20250313
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.32.1
vbuild==0.8.2
vila==0.5.0
Wand==0.6.13
wandb==0.19.8
wasabi==1.1.3
watchfiles==1.0.3
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
xxhash==3.5.0
yarl==1.18.3
ollama
